config file, and utils for functions needed everywhere.
