+++ config.js
@@ -1,0 +1,7 @@
+// List could expand
+const VOWELS = ['a', 'e', 'i', 'o', 'u', 
+                'A', 'E', 'I', 'O', 'U'];
+
+module.exports = {
+    VOWELS
+};
