+++ test.input
@@ -1,2 +1,3 @@
 zallin was here
+baker
 lol
